Refactor the top menu from "Edit recipe" page into EditRecipeTopMenu component.
